Move non-route files to lib/, convert var to const (mostly)
