+++ lib/authenticate.js
@@ -11,6 +11,10 @@ const scopes = [
   'https://www.googleapis.com/auth/calendar'
 ]
 
+google.options({
+  auth: oauth2Client
+})
+
 function getAccessToken (res) {
   const url = oauth2Client.generateAuthUrl({
     access_type: 'offline',
@@ -31,10 +35,6 @@ function setToken (tokenCode, res) {
     }
   })
 }
-
-google.options({
-  auth: oauth2Client
-})
 
 module.exports = {
   api: google,
