+++ lib/clean.js
@@ -1,8 +1,7 @@
-var authenticate = require('../lib/authenticate.js')
-var oauth2Client = authenticate.client
-var calendarAPI = authenticate.api.calendar('v3')
-var newCalendarId = null
-var debug = require('debug')('clean')
+const authenticate = require('../lib/authenticate.js')
+const oauth2Client = authenticate.client
+const calendarAPI = authenticate.api.calendar('v3')
+const debug = require('debug')('lib:clean')
 
 const FEB_20_START = '2017-02-20T00:00:00-05:00'
 const FEB_20_END = '2017-02-20T23:00:00-05:00'
