+++ lib/insert.js
@@ -1,10 +1,10 @@
-var async = require('async')
-var debug = require('debug')('insert')
-var authenticate = require('../lib/authenticate.js')
-var oauth2Client = authenticate.client
-var calendarAPI = authenticate.api.calendar('v3')
+const async = require('async')
+const debug = require('debug')('routes:insert')
+const authenticate = require('../lib/authenticate')
+const oauth2Client = authenticate.client
+const calendarAPI = authenticate.api.calendar('v3')
 
-var parser = require('./parse.js')
+const parser = require('./parse')
 
 const UNTIL_DATE = '20170503T120000Z'
