+++ lib/parse.js
@@ -1,4 +1,4 @@
-var debug = require('debug')('parse')
+const debug = require('debug')('lib:parse')
 
 const MON_START = '20170123'
 const TUE_START = '20170124'
