+++ package.json
@@ -17,6 +17,7 @@
     "googleapis": "^2.1.3",
     "htmlparser": "^1.7.7",
     "jade": "~1.11.0",
+    "moment": "^2.20.1",
     "morgan": "~1.6.1",
     "rotating-file-stream": "^1.3.4",
     "serve-favicon": "~2.3.0"
