+++ views/index.jade
@@ -1,6 +1,6 @@
 extends layout
 block content
-    h1(class="title")= title
-    h2(class="title")= subtitle
+    h1(class="title") BU Schedule to Google Calendar Import
+    h2(class="title") Use this tool to create a google calendar from your BU class schedule.
     button(onclick="goCalendar()" class="start").
       Begin
